[DR] CBL recover to no ratio version
- only participate or no participate (0/1)
- no ratio in HEMS & CEMS to multiply CBL
- in participate table some value is between 0 and 1, just ceil it

+++ include/SQLFunction.hpp
@@ -91,7 +91,7 @@ typedef struct
 	int minDecrease_power;
 	int customer_baseLine;
 	// LHEMS
-	int *participate_array;
+	float *participate_array;
 	float household_CBL;
 } DEMANDRESPONSE;
